kern.c: Update map declaration to have BTF support (#4)
Use the recent syntax specific to eBPF maps, and which allows for the
creation of BTF information related to the map in the object file. With
modern tooling this information can be loaded into the kernel when the
map is created, such that we can later retrieve it. For example, we can
get the following from bpftool:

    # bpftool map dump id 42
    [{
            "key": 340118,
            "value": 1
        }
    ]

Compare with the current version (no BTF):

    # bpftool map dump id 42
    key: 96 30 05 00 00 00 00 00  value: 01
    Found 1 element

BTF support for eBPF programs is already present through the "-g" flag
used to compile the program.